Criação do formulário para cadastro e correções diversas.

+++ lojinha-common/src/main/java/br/senac/tads/dsw/lojinha/common/service/fakeimpl/CategoriaServiceFakeImpl.java
@@ -36,17 +36,17 @@ import java.util.Map;
  */
 public class CategoriaServiceFakeImpl implements CategoriaService {
 
-  private static final Map<Long, Categoria> MAP_CATEGORIA = new LinkedHashMap<Long, Categoria>();
+  private static final Map<Integer, Categoria> MAP_CATEGORIA = new LinkedHashMap<Integer, Categoria>();
 
   static {
-    MAP_CATEGORIA.put(1L, new Categoria(1, "Bolo"));
-    MAP_CATEGORIA.put(2L, new Categoria(2, "Torta"));
-    MAP_CATEGORIA.put(3L, new Categoria(3, "Chocolate"));
-    MAP_CATEGORIA.put(4L, new Categoria(4, "Morango"));
-    MAP_CATEGORIA.put(5L, new Categoria(5, "Light"));
-    MAP_CATEGORIA.put(6L, new Categoria(6, "Crocante"));
-    MAP_CATEGORIA.put(7L, new Categoria(7, "Abacaxi"));
-    MAP_CATEGORIA.put(8L, new Categoria(8, "Coco"));
+    MAP_CATEGORIA.put(1, new Categoria(1, "Bolo"));
+    MAP_CATEGORIA.put(2, new Categoria(2, "Torta"));
+    MAP_CATEGORIA.put(3, new Categoria(3, "Chocolate"));
+    MAP_CATEGORIA.put(4, new Categoria(4, "Morango"));
+    MAP_CATEGORIA.put(5, new Categoria(5, "Light"));
+    MAP_CATEGORIA.put(6, new Categoria(6, "Crocante"));
+    MAP_CATEGORIA.put(7, new Categoria(7, "Abacaxi"));
+    MAP_CATEGORIA.put(8, new Categoria(8, "Coco"));
   }
 
   @Override
@@ -55,7 +55,7 @@ public class CategoriaServiceFakeImpl implements CategoriaService {
   }
 
   @Override
-  public Categoria obter(Long id) {
+  public Categoria obter(int id) {
     return MAP_CATEGORIA.get(id);
   }
 }
